fix: textfieldを共通のComponentとしてまとめました
