+++ lib/ui/screen/edit/edit_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:food_gram_app/ui/component/app_account_text_field.dart';
 import 'package:food_gram_app/ui/component/app_icon.dart';
 import 'package:food_gram_app/ui/component/app_loading.dart';
+import 'package:food_gram_app/ui/component/app_text_field.dart';
 import 'package:food_gram_app/ui/screen/edit/edit_view_model.dart';
 import 'package:food_gram_app/utils/provider/loading.dart';
 import 'package:go_router/go_router.dart';
@@ -100,7 +100,6 @@ class EditScreen extends ConsumerWidget {
                       SizedBox(height: 30),
                       Divider(),
                       AppNameTextField(
-                        title: 'ユーザー名',
                         controller: controller.nameTextController,
                       ),
                       Divider(),
